CharacterModels
Yo yo whats up dog. Character models for the Manmallow man, Cyclops and Rockus are in the "Carltons folder" folder yo.  They all move good but remember to use the global game objects dawg. Also don't look at the skeleton set up for the rockus, it looks fucked but it works fine and everything is connected how its meant to be. thats it from me, Peace.